Update mixer.ts
https://github.com/ConnorChristie/Audio-Mixer/issues/5#issuecomment-1345594399

Unfortunately it seems this repo is no longer maintained. I was able to fork the repo (to our company repo) - you will be able to fix the CPU issue by doing the following here: https://github.com/ConnorChristie/Audio-Mixer/blob/master/src/mixer.ts#L82

Replace this bit:

clearImmediate(this._timer)
this._timer = setImmediate(this._read.bind(this));
with

clearTimeout(this._timer)
this._timer = setTimeout(this._read.bind(this), 0);
I dont know what kind of impact it would have in terms of realtime processing. But for our use-case it works and all 2000 unit tests are still passing after this change. But CPU basically went down to 0 :)